requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.0.3
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.45.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
numpy==1.26.4
oauthlib==3.3.1
openai==2.13.0
opt_einsum==3.4.0
packaging==25.0
pillow==10.4.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.0.3
wrapt==2.0.1
